+++ src/lib.rs
@@ -8,7 +8,7 @@
 `glam` is built with SIMD in mind. Currently only SSE2 on x86/x86_64 is
 supported as this is what stable Rust supports.
 
-* Single precision float (`f32`) support only
+* Single precision float `f32` support only (for now)
 * SSE2 storage and optimization for many types, including `Mat2`, `Mat4`,
   `Quat`, `Vec3A` and `Vec4`
 * Scalar fallback implementations exist when SSE2 is not available
@@ -34,6 +34,47 @@ Matrices are stored in memory in column-major order.
 Some `glam` types use SIMD for storage meaning they are 16 byte aligned, these
 types include `Mat2`, `Mat4`, `Quat`, `Vec3A` and `Vec4`.
 
+When SSE2 is not available on the target architecture this type will still be 16
+byte aligned so that object sizes and layouts will not change between
+architectures.
+
+SIMD support can be disabled entirely using the `scalar-math` feature. This
+feature will also disable SIMD alignment meaning most types will use native
+`f32` alignment of 4 bytes.
+
+All the main `glam` types are `#[repr(C)]`, so they are possible to expose as
+struct members to C interfaces if desired. Be mindful of Vec3A's extra padding
+though.
+
+## Accessing internal data
+
+The SIMD types that `glam` builds on are opaque and their contents are not
+directly accessible. Because of this all types use getter and setter methods
+instead of providing direct access, regardless of whether they are using scalar
+or SIMD storage.
+
+```
+use glam::Vec4;
+let mut v = Vec4::new(1.0, 2.0, 3.0, 4.0);
+assert_eq!(v.y(), 2.0);
+v.set_z(1.0);
+assert_eq!(v.z(), 1.0);
+*v.x_mut() = 2.0;
+assert_eq!(v.x(), 2.0);
+```
+
+If you need to access multiple elements it is easier to convert the type to a
+tuple or array:
+
+```
+use glam::Vec4;
+let v = Vec4::new(1.0, 2.0, 3.0, 4.0);
+let (x, y, z, w) = v.into();
+assert_eq!((x, y, z, w), (1.0, 2.0, 3.0, 4.0));
+```
+
+## Vec3A
+
 `Vec3A` is a SIMD optimized version of the `Vec3` type, which due to 16 byte
 alignment results in `Vec3A` containing 4 bytes of padding making it 16 bytes
 in size in total.
@@ -47,42 +88,71 @@ Despite this wasted space the SIMD version tends to outperform the `f32`
 implementation in [**mathbench**](https://github.com/bitshifter/mathbench-rs)
 benchmarks.
 
-When SSE2 is not available on the target architecture this type will still be 16
-byte aligned, so object sizes and layouts will not change between architectures.
+`glam` treats `Vec3` as the default vector 3 type and `Vec3A` a special case for
+optimization. When methods need to return a vector 3 type they will generally
+return `Vec3`.
 
-SIMD support can be disabled entirely using the `scalar-math` feature. This
-feature will also disable SIMD alignment meaning most types will use native
-`f32` alignment of 4 bytes.
-
-All the main `glam` types are tagged with #[repr(C)], so they are possible
-to expose as struct members to C interfaces if desired. Be mindful of Vec3A's
-extra padding though.
-
-## Accessing internal data
-
-The SIMD types that `glam` builds on are opaque and their contents are not
-directly accessible. Because of this all types use getter and setter methods
-instead of providing direct access, regardless of whether they are using scalar
-or SIMD storage.
+There are `From` trait implementations for converting from `Vec4` to a `Vec3A`
+and between `Vec3` and `Vec3A` (and vice versa).
 
 ```
-use glam::Vec3A;
-let mut v = Vec3A::new(1.0, 2.0, 3.0);
-assert_eq!(v.y(), 2.0);
-v.set_z(1.0);
-assert_eq!(v.z(), 1.0);
-*v.x_mut() = 2.0;
-assert_eq!(v.x(), 2.0);
+use glam::{Vec3, Vec3A, Vec4};
+
+let v4 = Vec4::new(1.0, 2.0, 3.0, 4.0);
+
+// Convert from `Vec4` to `Vec3A`, this is a no-op if SIMD is supported.
+let v3a = Vec3A::from(v4);
+assert_eq!(Vec3A::new(1.0, 2.0, 3.0), v3a);
+
+// Convert from `Vec3A` to `Vec3`.
+let v3 = Vec3::from(v3a);
+assert_eq!(Vec3::new(1.0, 2.0, 3.0), v3);
+
+// Convert from `Vec3` to `Vec3A`.
+let v3a = Vec3A::from(v3);
+assert_eq!(Vec3A::new(1.0, 2.0, 3.0), v3a);
 ```
 
-If you need to access multiple elements it is easier to convert the type to a
-tuple or array:
+## Vector swizzles
+
+`glam` vector types have functions allowing elements of vectors to be reordered,
+this includes creating a vector of a different size from the vectors elements.
+
+The swizzle functions are implemented using traits to add them to each vector
+type. This is primarily because there are a lot of swizzle functions which can
+obfuscate the other vector functions in documentation and so on. The traits are
+`Vec2Swizzles`, `Vec3Swizzles`, `Vec3ASwizzles` and `Vec4Swizzles`.
+
+Note that `Vec2Swizzles`, `Vec3Swizzles` and `Vec4Swizzles` will return a `Vec3`
+type for 3 element swizzles. `Vec3ASwizzles` will return a `Vec3A` for 3 element
+swizzles.
 
 ```
-use glam::Vec3A;
-let v = Vec3A::new(1.0, 2.0, 3.0);
-let (x, y, z) = v.into();
-assert_eq!((x, y, z), (1.0, 2.0, 3.0));
+use glam::{Vec2, Vec2Swizzles, Vec3, Vec3Swizzles, Vec3A, Vec3ASwizzles, Vec4, Vec4Swizzles};
+
+let v = Vec4::new(1.0, 2.0, 3.0, 4.0);
+
+// Reverse elements of `v`, if SIMD is supported this will use a vector shuffle.
+let wzyx = v.wzyx();
+assert_eq!(Vec4::new(4.0, 3.0, 2.0, 1.0), wzyx);
+
+// Swizzle the yzw elements of `v` into a `Vec3`
+let yzw = v.yzw();
+assert_eq!(Vec3::new(2.0, 3.0, 4.0), yzw);
+
+// To swizzle a `Vec4` into a `Vec3A` swizzle the `Vec4` first then convert to
+// `Vec3A`. If SIMD is supported this will use a vector shuffle. The last
+// element of the shuffled `Vec4` is ignored by the `Vec3A`.
+let yzw = Vec3A::from(v.yzwx());
+assert_eq!(Vec3A::new(2.0, 3.0, 4.0), yzw);
+
+// You can swizzle from a `Vec4` to a `Vec2`
+let xy = v.xy();
+assert_eq!(Vec2::new(1.0, 2.0), xy);
+
+// And back again
+let yyxx = xy.yyxx();
+assert_eq!(Vec4::new(2.0, 2.0, 1.0, 1.0), yyxx);
 ```
 
 ## SIMD and scalar consistency
@@ -95,9 +165,9 @@ The SIMD versions implement the `core::fmt::Debug` and `core::fmt::Display`
 traits so they print the same as the scalar version.
 
 ```
-use glam::Vec3A;
-let a = Vec3A::new(1.0, 2.0, 3.0);
-assert_eq!(format!("{}", a), "[1, 2, 3]");
+use glam::Vec4;
+let a = Vec4::new(1.0, 2.0, 3.0, 4.0);
+assert_eq!(format!("{}", a), "[1, 2, 3, 4]");
 ```
 
 ## Feature gates
@@ -125,8 +195,9 @@ mod macros;
 pub mod f32;
 
 pub use self::f32::{
-    mat2, mat3, mat4, quat, vec2, vec3, vec3a, vec4, Mat2, Mat3, Mat4, Quat, Vec2, Vec2Mask, Vec3,
-    Vec3A, Vec3AMask, Vec3Mask, Vec4, Vec4Mask,
+    mat2, mat3, mat4, quat, vec2, vec3, vec3a, vec4, Mat2, Mat3, Mat4, Quat, Vec2, Vec2Mask,
+    Vec2Swizzles, Vec3, Vec3A, Vec3AMask, Vec3ASwizzles, Vec3Mask, Vec3Swizzles, Vec4, Vec4Mask,
+    Vec4Swizzles,
 };
 
 #[cfg(feature = "transform-types")]